fix: move useOutsideClick
